+++ internal/applications/filesharer/pkg/app/app.go
@@ -36,7 +36,7 @@ type sApp struct {
 	fStgPath string
 
 	fIntServiceHTTP *http.Server
-	fIncServiceHTTP *http.Server
+	fExtServiceHTTP *http.Server
 	fServicePPROF   *http.Server
 
 	fHTTPLogger logger.ILogger
@@ -62,8 +62,8 @@ func NewApp(
 func (p *sApp) Run(pCtx context.Context) error {
 	services := []internal_types.IServiceF{
 		p.runListenerPPROF,
-		p.runIncomingListenerHTTP,
-		p.runInterfaceListenerHTTP,
+		p.runExternalListenerHTTP,
+		p.runInternalListenerHTTP,
 	}
 
 	ctx, cancel := context.WithCancel(pCtx)
@@ -105,8 +105,8 @@ func (p *sApp) enable(pCtx context.Context) state.IStateF {
 		)
 
 		p.initServicePPROF()
-		p.initIncomingServiceHTTP(pCtx, hlsClient)
-		p.initInterfaceServiceHTTP(pCtx, hlsClient)
+		p.initExternalServiceHTTP(pCtx, hlsClient)
+		p.initInternalServiceHTTP(pCtx, hlsClient)
 
 		p.fStdfLogger.PushInfo(fmt.Sprintf(
 			"%s is started; %s",
@@ -132,6 +132,7 @@ func (p *sApp) disable(pCancel context.CancelFunc, pWg *sync.WaitGroup) state.IS
 
 func (p *sApp) runListenerPPROF(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
+	defer func() { <-pCtx.Done() }()
 
 	if p.fConfig.GetAddress().GetPPROF() == "" {
 		return
@@ -144,12 +145,28 @@ func (p *sApp) runListenerPPROF(pCtx context.Context, wg *sync.WaitGroup, pChErr
 			return
 		}
 	}()
-
-	<-pCtx.Done()
 }
 
-func (p *sApp) runInterfaceListenerHTTP(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runExternalListenerHTTP(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
+	defer func() { <-pCtx.Done() }()
+
+	if p.fConfig.GetAddress().GetExternal() == "" {
+		return
+	}
+
+	go func() {
+		err := p.fExtServiceHTTP.ListenAndServe()
+		if err != nil && !errors.Is(err, http.ErrServerClosed) {
+			pChErr <- err
+			return
+		}
+	}()
+}
+
+func (p *sApp) runInternalListenerHTTP(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+	defer wg.Done()
+	defer func() { <-pCtx.Done() }()
 
 	go func() {
 		err := p.fIntServiceHTTP.ListenAndServe()
@@ -158,28 +175,12 @@ func (p *sApp) runInterfaceListenerHTTP(pCtx context.Context, wg *sync.WaitGroup
 			return
 		}
 	}()
-
-	<-pCtx.Done()
-}
-
-func (p *sApp) runIncomingListenerHTTP(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
-	defer wg.Done()
-
-	go func() {
-		err := p.fIncServiceHTTP.ListenAndServe()
-		if err != nil && !errors.Is(err, http.ErrServerClosed) {
-			pChErr <- err
-			return
-		}
-	}()
-
-	<-pCtx.Done()
 }
 
 func (p *sApp) stop() error {
 	err := closer.CloseAll([]io.Closer{
 		p.fIntServiceHTTP,
-		p.fIncServiceHTTP,
+		p.fExtServiceHTTP,
 		p.fServicePPROF,
 	})
 	if err != nil {