+++ internal/applications/filesharer/pkg/app/app_test.go
@@ -59,8 +59,8 @@ func TestApp(t *testing.T) {
 			FPageOffset: 10,
 		},
 		FAddress: &config.SAddress{
-			FInterface: testutils.TgAddrs[37],
-			FIncoming:  testutils.TgAddrs[29],
+			FInternal: testutils.TgAddrs[37],
+			FExternal: testutils.TgAddrs[29],
 		},
 		FConnection: "test_connection",
 	})