+++ internal/applications/filesharer/pkg/app/config/types.go
@@ -28,7 +28,7 @@ type IConfigSettings interface {
 }
 
 type IAddress interface {
-	GetInterface() string
-	GetIncoming() string
+	GetInternal() string
+	GetExternal() string
 	GetPPROF() string
 }