+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - `internal/service/pkg/config`: rename GetLimitMessageSizeBytes -> GetPayloadSizeBytes
 - `internal/utils/logger/anon`: addr, conn -> variable params
+- `internal/applications/*`: rename config params incoming,interface -> external,internal
+- `internal/applications/filesharer`: external param is now can be omitempty
 
 ### BUG FIXES